demos/voice_AI_pi_led/pi_api_demo/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
RPi.GPIO>=0.7.1
pydantic>=2.0.0
